+++ src/main/java/edu/miu/cs544/sujan/jobboot/service/JobService.java
@@ -5,13 +5,13 @@ import edu.miu.cs544.sujan.jobboot.entity.Job;
 import java.util.List;
 
 public interface JobService {
-    Job getJobById(int id);
+    Job getJobById(Long id);
 
     List<Job> getJobs();
 
     Job createJob(Job job);
 
-    void deleteJob(int id);
+    void deleteJob(Long id);
 
-    Job updateJob(int id, Job job);
+    Job updateJob(Long id, Job job);
 }
